requirements.txt
alembic==1.11.1
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
libcst==0.4.9
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-extensions==1.0.0
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0
rich==13.3.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
strawberry-graphql==0.177.3
typing-inspect==0.8.0
typing_extensions==4.5.0
uvicorn==0.21.1
